Move network graph to submodule
