+++ src/main/network/mod.rs
@@ -1,5 +1,4 @@
 pub mod graph;
 mod packet;
-mod petgraph_wrapper;
 mod relay;
 pub mod router;
